vinayak/requirements.txt
opencv-python
mediapipe
ultralytics
tensorflow-cpu==2.12.0
openvino
ipython
imageio
flask
flask_cors